requirements.txt
Django==4.2.17
djangorestframework==3.15.2
drf_yasg==1.20.0
djangorestframework-simplejwt
python-telegram-bot
requests
drf-spectacular
django-cors-headers==4.6.0
gunicorn==22.0.0
Pillow==11.1.0
